Harden codec streams and implement LCW compression
Corrupt or crafted data could overrun buffers in the compression pipes and
straws, the WSA reader and the MIX reader:

- LCW, LZW and LZO pipes and straws checked no block header. Counts must
  now fit the buffer and decode to the declared size; a corrupt block stops
  the stream after the valid blocks before it. LCW gets a bounded decoder,
  LZW rejects codes past its dictionary, and LZO uses a new
  lzo1x_decompress_safe that checks every read, write and back-reference.
- LCW_Comp was a stub returning 0, so map and overlay packs written through
  it were empty. It is now a real encoder for the format the decoder reads,
  and the LCW and LZW buffer margins cover the worst-case output.
- WSA cleared only size % 65536 bytes of a caller's buffer and trusted frame
  sizes for frame 0 and later deltas.
- MIX headers with negative or oversized counts and sizes threw or
  over-allocated; Open now fails.
- LZOStraw's compress dictionary was too small for 64-bit pointers.
- The three-argument XMP_Encode looped ~4 billion times for a short buffer.

Also: the file seek wrappers take and return int64_t, and TD's keyframe
decoder loses dead 16-bit segment logic. New tests feed corrupt streams and
round-trip LCW data; they fail or crash on the previous code.

+++ src/sdllib/file.h
@@ -67,8 +67,11 @@ bool IO_Read_File(void* handle, void* buffer, size_t count,
 bool IO_Write_File(void* handle, const void* buffer, size_t count,
                    size_t& actual_written);
 
-size_t IO_Seek_File(void* handle, size_t offset, int origin);
-size_t IO_Get_File_Size(void* handle);
+// Seeks like fseek and returns the new position, or -1 on error. offset may
+// be negative for SEEK_CUR and SEEK_END.
+int64_t IO_Seek_File(void* handle, int64_t offset, int origin);
+// Returns the file's length in bytes, or -1 on error. Preserves the position.
+int64_t IO_Get_File_Size(void* handle);
 
 bool IO_Delete_File(const char* filename);
